code-review-app/requirements.txt
# Core MCP dependencies
mcp[cli]>=1.9.2

# Web server
uvicorn>=0.32.0

# HTTP client
httpx>=0.28.0

# Environment management
python-dotenv>=1.0.0

# OpenAI API (optional for future AI-powered analysis)
openai>=1.12.0

# Testing dependencies (dev)
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Code quality (dev)
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
